refactor: move Tasks features to domain folder
