+++ src/domains/tasks/features/Tasks/TasksDashboard.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 
-import BodyStyles from "../../components/styles/BodyStyles";
+import BodyStyles from "../../../../components/styles/BodyStyles";
 import { FilterTasks, TaskCategory, TasksList, TaskStatus } from "./TaskList";
 import axios from "axios";
 import { useSession } from "next-auth/react";
