+++ src/pages/mytasks/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import TasksDashboard from "../../features/Tasks/TasksDashboard";
+import TasksDashboard from "@Tasks/features/Tasks/TasksDashboard";
 
 const MyTasksPage: NextPage = () => {
   return <TasksDashboard myTasks={true} />;
