+++ src/pages/offers.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import { Dashboard } from "../features/Offers";
+import { Dashboard } from "@Tasks/features/Offers";
 
 const OffersPage: NextPage = () => {
   return <Dashboard />;
