+++ src/pages/tasks/createtask.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import CreateTask from "../../features/Tasks/CreateTask/CreateTask";
+import CreateTask from "@Tasks/features/Tasks/CreateTask/CreateTask";
 
 const CreateTaskPage: NextPage = () => {
   return <CreateTask />;
