+++ src/pages/tasks/index.tsx
@@ -1,7 +1,6 @@
-import type { NextPage } from "next";
-import TasksDashboard from "../../features/Tasks/TasksDashboard";
+import TasksDashboard from "@Tasks/features/Tasks/TasksDashboard";
 
-const TasksPage: NextPage = () => {
+const TasksPage = () => {
   return <TasksDashboard myTasks={false} />;
 };
